+++ tests/unit/test_transactions.py
@@ -1,5 +1,7 @@
 import unittest
 from pkmntypes import *
+import util
+from hypothesis import given, reproduce_failure, strategies as st
 
 
 class TestTransactions(unittest.TestCase):
@@ -159,6 +161,91 @@ class TestTransactions(unittest.TestCase):
 		self.assertEqual(
 			transaction.pretty(), "Altaria [Dragon] [Flying] took 14 damage."
 		)
+
+
+garbage_transactions = st.builds(
+	Transaction,
+	type=st.integers(),
+	name=st.text(min_size=1),
+	args=st.dictionaries(
+		st.text(),
+		st.one_of(
+			st.integers(), st.text(),
+			st.dictionaries(
+				st.text(),
+				st.one_of(st.integers(), st.text()),
+			)
+		),
+	)
+)
+
+# yapf: disable
+sampled_transactions = st.sampled_from(
+	[
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Pichu', 'NatDex': 172, 'Level': 30, 'Ability': 98, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [0, 1, 2, 4, 0, 2], 'EVs': [13, 18, 17, 22, 32, 21], 'Nature': 8, 'Stats': [52, 30, 15, 28, 28, 43], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 52, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 433, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Trick Room', 'Type': 8192, 'Category': 0, 'Targets': 12, 'Priority': -7, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 2048, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 225, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Dragon Breath', 'Type': 32768, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 60, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 30, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 232, 'CurrentPP': 35, 'MaxPP': 35, 'Name': 'Metal Claw', 'Type': 256, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 95, 'InitialMaxPP': 35, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18448, 'AffectedStat': 1, 'StatStages': 1}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 4096}}, 'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [2, 4, 3, 1, 1, 4], 'EVs': [18, 9, 7, 24, 41, 11], 'Nature': 8, 'Stats': [77, 54, 72, 37, 56, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 67, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 18, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Whirlwind', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': -6, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 264193, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 242, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Crunch', 'Type': 65536, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 20, 'Flags': 18452, 'AffectedStat': 2, 'StatStages': -1}, {'Id': 61, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bubble Beam', 'Type': 1024, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 65, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18432, 'AffectedStat': 5, 'StatStages': -1}, {'Id': 457, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Head Smash', 'Type': 32, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 150, 'Accuracy': 80, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -50, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}}, 'Move': {'Id': 225, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Dragon Breath', 'Type': 32768, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 60, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 30, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={ 'Move': {'Id': 225,'CurrentPP': 19,'MaxPP': 20,'Name': 'Dragon Breath','Type': 32768,'Category': 2,'Targets': 10,'Priority': 0,'Power': 60,'Accuracy': 100,'InitialMaxPP': 20,'MinHits': 0,'MaxHits': 0,'MinTurns': 0,'MaxTurns': 0,'Drain': 0,'Healing': 0,'CritRate': 0,'AilmentChance': 30,'FlinchChance': 0,'StatChance': 0,'Flags': 18432,'AffectedStat': 0,'StatStages': 0},'Amount': -1}),
+	Transaction(type=1, name="DamageTransaction", args={'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [2, 4, 3, 1, 1, 4], 'EVs': [18, 9, 7, 24, 41, 11], 'Nature': 8, 'Stats': [77, 54, 72, 37, 56, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 67, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 18, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Whirlwind', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': -6, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 264193, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 242, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Crunch', 'Type': 65536, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 20, 'Flags': 18452, 'AffectedStat': 2, 'StatStages': -1}, {'Id': 61, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bubble Beam', 'Type': 1024, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 65, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18432, 'AffectedStat': 5, 'StatStages': -1}, {'Id': 457, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Head Smash', 'Type': 32, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 150, 'Accuracy': 80, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -50, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}}, 'Move': {'Id': 225, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Dragon Breath', 'Type': 32768, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 60, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 30, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, 'Damage': 10, 'StatusEffect': 0}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 1, 'Slot': 0, 'Team': 1, 'Pokemon': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [2, 4, 3, 1, 1, 4], 'EVs': [18, 9, 7, 24, 41, 11], 'Nature': 8, 'Stats': [77, 54, 72, 37, 56, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 67, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 18, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Whirlwind', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': -6, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 264193, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 242, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Crunch', 'Type': 65536, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 20, 'Flags': 18452, 'AffectedStat': 2, 'StatStages': -1}, {'Id': 61, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bubble Beam', 'Type': 1024, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 65, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18432, 'AffectedStat': 5, 'StatStages': -1}, {'Id': 457, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Head Smash', 'Type': 32, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 150, 'Accuracy': 80, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -50, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}}, 'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Pichu', 'NatDex': 172, 'Level': 30, 'Ability': 98, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [0, 1, 2, 4, 0, 2], 'EVs': [13, 18, 17, 22, 32, 21], 'Nature': 8, 'Stats': [52, 30, 15, 28, 28, 43], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 52, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 433, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Trick Room', 'Type': 8192, 'Category': 0, 'Targets': 12, 'Priority': -7, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 2048, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 225, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Dragon Breath', 'Type': 32768, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 60, 'Accuracy': 100, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 30, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 232, 'CurrentPP': 35, 'MaxPP': 35, 'Name': 'Metal Claw', 'Type': 256, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 95, 'InitialMaxPP': 35, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18448, 'AffectedStat': 1, 'StatStages': 1}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 4096}}, 'Move': {'Id': 18, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Whirlwind', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': -6, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 264193, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 18, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Whirlwind', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': -6, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 264193, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 1, 'Slot': 0, 'Team': 1, 'Pokemon': {'Name': 'Mantyke', 'NatDex': 458, 'Level': 30, 'Ability': 120, 'TotalExperience': 33750, 'Gender': 0, 'IVs': [4, 0, 0, 3, 0, 0], 'EVs': [21, 10, 36, 24, 5, 0], 'Nature': 8, 'Stats': [69, 17, 37, 43, 77, 35], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 69, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 293, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Camouflage', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 164, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Substitute', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 528384, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 283, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Endeavor', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 207, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Swagger', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 1, 'StatStages': 2}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 1028}}, 'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [5, 2, 2, 5, 2, 4], 'EVs': [19, 40, 40, 23, 31, 3], 'Nature': 8, 'Stats': [78, 56, 74, 38, 55, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 78, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 438, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Power Whip', 'Type': 2048, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 85, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 463, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Magma Storm', 'Type': 512, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 100, 'Accuracy': 75, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 5, 'MaxTurns': 6, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 100, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 173, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Snore', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 30, 'StatChance': 0, 'Flags': 1067009, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 412, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Energy Ball', 'Type': 2048, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18434, 'AffectedStat': 4, 'StatStages': -1}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}}, 'Move': {'Id': 293, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Camouflage', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 293, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Camouflage', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [5, 2, 2, 5, 2, 4], 'EVs': [19, 40, 40, 23, 31, 3], 'Nature': 8, 'Stats': [78, 56, 74, 38, 55, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 78, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 438, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Power Whip', 'Type': 2048, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 85, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 463, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Magma Storm', 'Type': 512, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 100, 'Accuracy': 75, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 5, 'MaxTurns': 6, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 100, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 173, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Snore', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 30, 'StatChance': 0, 'Flags': 1067009, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 412, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Energy Ball', 'Type': 2048, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18434, 'AffectedStat': 4, 'StatStages': -1}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}}, 'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Mantyke', 'NatDex': 458, 'Level': 30, 'Ability': 120, 'TotalExperience': 33750, 'Gender': 0, 'IVs': [4, 0, 0, 3, 0, 0], 'EVs': [21, 10, 36, 24, 5, 0], 'Nature': 8, 'Stats': [69, 17, 37, 43, 77, 35], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 69, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 293, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Camouflage', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 164, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Substitute', 'Type': 1, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 528384, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 283, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Endeavor', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 207, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Swagger', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 1, 'StatStages': 2}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 1028}}, 'Move': {'Id': 463, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Magma Storm', 'Type': 512, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 100, 'Accuracy': 75, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 5, 'MaxTurns': 6, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 100, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 463, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Magma Storm', 'Type': 512, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 100, 'Accuracy': 75, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 5, 'MaxTurns': 6, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 100, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=15, name="MoveFailTransaction", args={'User': {'Name': 'Marowak', 'NatDex': 105, 'Level': 30, 'Ability': 49, 'TotalExperience': 27000, 'Gender': 0, 'IVs': [5, 2, 2, 5, 2, 4], 'EVs': [19, 40, 40, 23, 31, 3], 'Nature': 8, 'Stats': [78, 56, 74, 38, 55, 33], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 78, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 438, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Power Whip', 'Type': 2048, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 85, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 463, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Magma Storm', 'Type': 512, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 100, 'Accuracy': 75, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 5, 'MaxTurns': 6, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 100, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 173, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Snore', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 30, 'StatChance': 0, 'Flags': 1067009, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 412, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Energy Ball', 'Type': 2048, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 10, 'Flags': 18434, 'AffectedStat': 4, 'StatStages': -1}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 16}, 'Reason': 1}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 66, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Lileep', 'NatDex': 345, 'Level': 30, 'Ability': 104, 'TotalExperience': 37800, 'Gender': 0, 'IVs': [4, 1, 3, 0, 3, 1], 'EVs': [26, 33, 13, 8, 26, 24], 'Nature': 8, 'Stats': [82, 32, 53, 42, 59, 20], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 82, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 265, 'CurrentPP': 9, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 432, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Defog', 'Type': 4, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 166, 'CurrentPP': 1, 'MaxPP': 1, 'Name': 'Sketch', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 1, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 474, 'CurrentPP': 0, 'MaxPP': 0, 'Name': 'Shadow End', 'Type': 0, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 60, 'InitialMaxPP': 0, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 2080}}, 'Move': {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 1, 'Slot': 0, 'Team': 1, 'Pokemon': {'Name': 'Lileep', 'NatDex': 345, 'Level': 30, 'Ability': 104, 'TotalExperience': 37800, 'Gender': 0, 'IVs': [4, 1, 3, 0, 3, 1], 'EVs': [26, 33, 13, 8, 26, 24], 'Nature': 8, 'Stats': [82, 32, 53, 42, 59, 20], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 82, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 265, 'CurrentPP': 9, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 432, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Defog', 'Type': 4, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 166, 'CurrentPP': 1, 'MaxPP': 1, 'Name': 'Sketch', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 1, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 474, 'CurrentPP': 0, 'MaxPP': 0, 'Name': 'Shadow End', 'Type': 0, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 60, 'InitialMaxPP': 0, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 2080}}, 'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 66, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Move': {'Id': 265, 'CurrentPP': 9, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 265, 'CurrentPP': 9, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=1, name="DamageTransaction", args={'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 66, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Move': {'Id': 265, 'CurrentPP': 9, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Damage': 16, 'StatusEffect': 0}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 1, 'Slot': 0, 'Team': 1, 'Pokemon': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}}, 'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Solrock', 'NatDex': 338, 'Level': 30, 'Ability': 48, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 1, 4, 2, 3], 'EVs': [15, 7, 31, 5, 31, 11], 'Nature': 8, 'Stats': [94, 62, 58, 39, 46, 48], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 94, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 198, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Bone Rush', 'Type': 16, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 25, 'Accuracy': 90, 'InitialMaxPP': 10, 'MinHits': 2, 'MaxHits': 5, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 339, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bulk Up', 'Type': 2, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 2, 'StatStages': 1}, {'Id': 13, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Razor Wind', 'Type': 1, 'Category': 2, 'Targets': 11, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 1, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18440, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 8224}}, 'Move': {'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, 'Amount': -1}),
+	Transaction(type=16, name="ModifyStatTransaction", args={'Target': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}, 'SelfInflicted': True, 'Stat': 5, 'Stages': 2}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Solrock', 'NatDex': 338, 'Level': 30, 'Ability': 48, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 1, 4, 2, 3], 'EVs': [15, 7, 31, 5, 31, 11], 'Nature': 8, 'Stats': [94, 62, 58, 39, 46, 48], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 94, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 198, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Bone Rush', 'Type': 16, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 25, 'Accuracy': 90, 'InitialMaxPP': 10, 'MinHits': 2, 'MaxHits': 5, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 339, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bulk Up', 'Type': 2, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 2, 'StatStages': 1}, {'Id': 13, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Razor Wind', 'Type': 1, 'Category': 2, 'Targets': 11, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 1, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18440, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 8224}}, 'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}}, 'Move': {'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=1, name="DamageTransaction", args={'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}}, 'Move': {'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Damage': 75, 'StatusEffect': 0}),
+	Transaction(type=10, name="FaintTransaction", args={'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}}}),
+	Transaction(type=2, name="FriendshipTransaction", args={'Target': {'Name': 'Delibird', 'NatDex': 225, 'Level': 30, 'Ability': 35, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 0, 1, 0, 3], 'EVs': [13, 2, 27, 31, 33, 17], 'Nature': 8, 'Stats': [67, 38, 33, 46, 34, 52], 'StatModifiers': [0, 0, 0, 0, 0, 2, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 0, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 397, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Rock Polish', 'Type': 32, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 5, 'StatStages': 2}, {'Id': 193, 'CurrentPP': 40, 'MaxPP': 40, 'Name': 'Foresight', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 40, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 134, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Kinesis', 'Type': 8192, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 80, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 7, 'StatStages': -1}, {'Id': 227, 'CurrentPP': 5, 'MaxPP': 5, 'Name': 'Encore', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 281601, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': -1, 'OriginalTrainerID': 0, 'Type': 16388}, 'Amount': -1}),
+	Transaction(type=3, name="EVTransaction", args={'Target': {'Name': 'Solrock', 'NatDex': 338, 'Level': 30, 'Ability': 48, 'TotalExperience': 21600, 'Gender': 0, 'IVs': [0, 1, 1, 4, 2, 3], 'EVs': [15, 7, 31, 5, 31, 12], 'Nature': 8, 'Stats': [94, 62, 58, 39, 46, 48], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 94, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 387, 'CurrentPP': 4, 'MaxPP': 5, 'Name': 'Last Resort', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 140, 'Accuracy': 100, 'InitialMaxPP': 5, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 198, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Bone Rush', 'Type': 16, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 25, 'Accuracy': 90, 'InitialMaxPP': 10, 'MinHits': 2, 'MaxHits': 5, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18432, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 339, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Bulk Up', 'Type': 2, 'Category': 0, 'Targets': 7, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 524288, 'AffectedStat': 2, 'StatStages': 1}, {'Id': 13, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Razor Wind', 'Type': 1, 'Category': 2, 'Targets': 11, 'Priority': 0, 'Power': 80, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 1, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18440, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 8224}, 'Stat': 5, 'Amount': 1}),
+	Transaction(type=13, name="EndBattleTransaction", args={'Reason': 0, 'Winner': 0}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 50, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Target': {'Party': 1, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Lileep', 'NatDex': 345, 'Level': 30, 'Ability': 104, 'TotalExperience': 37800, 'Gender': 0, 'IVs': [4, 1, 3, 0, 3, 1], 'EVs': [26, 33, 13, 8, 26, 24], 'Nature': 8, 'Stats': [82, 32, 53, 42, 59, 20], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 82, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 265, 'CurrentPP': 8, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 432, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Defog', 'Type': 4, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 166, 'CurrentPP': 1, 'MaxPP': 1, 'Name': 'Sketch', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 1, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 474, 'CurrentPP': 0, 'MaxPP': 0, 'Name': 'Shadow End', 'Type': 0, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 60, 'InitialMaxPP': 0, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 2080}}, 'Move': {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=15, name="MoveFailTransaction", args={'User': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 50, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}, 'Reason': 1}),
+	Transaction(type=0, name="UseMoveTransaction", args={'User': {'Party': 1, 'Slot': 0, 'Team': 1, 'Pokemon': {'Name': 'Lileep', 'NatDex': 345, 'Level': 30, 'Ability': 104, 'TotalExperience': 37800, 'Gender': 0, 'IVs': [4, 1, 3, 0, 3, 1], 'EVs': [26, 33, 13, 8, 26, 24], 'Nature': 8, 'Stats': [82, 32, 53, 42, 59, 20], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 82, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 265, 'CurrentPP': 8, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 432, 'CurrentPP': 15, 'MaxPP': 15, 'Name': 'Defog', 'Type': 4, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280577, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 166, 'CurrentPP': 1, 'MaxPP': 1, 'Name': 'Sketch', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 0, 'InitialMaxPP': 1, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 474, 'CurrentPP': 0, 'MaxPP': 0, 'Name': 'Shadow End', 'Type': 0, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 120, 'Accuracy': 60, 'InitialMaxPP': 0, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 2080}}, 'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 50, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Move': {'Id': 265, 'CurrentPP': 8, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}}),
+	Transaction(type=5, name="PPTransaction", args={'Move': {'Id': 265, 'CurrentPP': 8, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Amount': -1}),
+	Transaction(type=1, name="DamageTransaction", args={'Target': {'Party': 0, 'Slot': 0, 'Team': 0, 'Pokemon': {'Name': 'Absol', 'NatDex': 359, 'Level': 30, 'Ability': 105, 'TotalExperience': 21760, 'Gender': 0, 'IVs': [0, 1, 1, 2, 0, 1], 'EVs': [41, 37, 21, 24, 2, 34], 'Nature': 8, 'Stats': [82, 86, 42, 52, 41, 52], 'StatModifiers': [0, 0, 0, 0, 0, 0, 0, 0, 0], 'StatusEffects': 0, 'CurrentHP': 50, 'HeldItem': {'Id': 0, 'Name': '', 'Category': 0, 'FlingPower': 0, 'FlingEffect': 0, 'Flags': 0}, 'Moves': [{'Id': 36, 'CurrentPP': 20, 'MaxPP': 20, 'Name': 'Take Down', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 90, 'Accuracy': 85, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': -25, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 261, 'CurrentPP': 14, 'MaxPP': 15, 'Name': 'Will-O-Wisp', 'Type': 512, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 85, 'InitialMaxPP': 15, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 280576, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 311, 'CurrentPP': 10, 'MaxPP': 10, 'Name': 'Weather Ball', 'Type': 1, 'Category': 2, 'Targets': 10, 'Priority': 0, 'Power': 50, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18434, 'AffectedStat': 0, 'StatStages': 0}, {'Id': 48, 'CurrentPP': 19, 'MaxPP': 20, 'Name': 'Supersonic', 'Type': 1, 'Category': 0, 'Targets': 10, 'Priority': 0, 'Power': 0, 'Accuracy': 55, 'InitialMaxPP': 20, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 2, 'MaxTurns': 5, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 1329153, 'AffectedStat': 0, 'StatStages': 0}], 'Friendship': 0, 'OriginalTrainerID': 0, 'Type': 65536}}, 'Move': {'Id': 265, 'CurrentPP': 8, 'MaxPP': 10, 'Name': 'Smelling Salts', 'Type': 1, 'Category': 1, 'Targets': 10, 'Priority': 0, 'Power': 70, 'Accuracy': 100, 'InitialMaxPP': 10, 'MinHits': 0, 'MaxHits': 0, 'MinTurns': 0, 'MaxTurns': 0, 'Drain': 0, 'Healing': 0, 'CritRate': 0, 'AilmentChance': 0, 'FlinchChance': 0, 'StatChance': 0, 'Flags': 18448, 'AffectedStat': 0, 'StatStages': 0}, 'Damage': 16, 'StatusEffect': 0}),
+	Transaction(type=12, name="WeatherTransaction", args={'Weather': 1, 'Turns': 5}),
+	]
+)
+# yapf: enable
+
+
+class TestPrettifyAllTransactions(unittest.TestCase):
+
+	@given(st.lists(st.one_of(garbage_transactions, sampled_transactions)))
+	def test_prettify_all_transactions_text_should_remain_in_bounds(self, transactions):
+		pretty_texts = util.prettify_all_transactions(transactions)
+		for text in pretty_texts:
+			self.assertLessEqual(len(text), 2048)
+
+	@given(st.lists(st.one_of(garbage_transactions, sampled_transactions)))
+	def test_prettify_all_transactions_should_never_contain_empty_string(
+		self, transactions
+	):
+		pretty_texts = util.prettify_all_transactions(transactions)
+		for text in pretty_texts:
+			self.assertIsNotNone(text)
+			self.assertIsInstance(text, str)
+			self.assertNotEqual(text, "")
+
+	@given(st.lists(sampled_transactions))
+	def test_prettify_all_transactions_should_render_all_transactions(self, transactions):
+		pretty_texts = util.prettify_all_transactions(transactions)
+		for text in pretty_texts:
+			self.assertFalse(text.startswith("TODO:"))
+			self.assertFalse(text.startswith("Failed:"))
+
+	def test_prettify_all_transactions_empty(self):
+		self.assertListEqual(util.prettify_all_transactions([]), [])
 
 
 if __name__ == "__main__":
